refactor(bayn): split evidence and cycle persistence
