+++ services/bayn/src/db/cycle-store/index.ts
@@ -1,10 +1,10 @@
 export {
   CycleStore,
   CycleStoreError,
-  CycleStoreLive,
   type CycleAcquireReceipt,
   type CycleAuthoritySlot,
   type CycleMutationReceipt,
   type CycleRecoveryScope,
   type CycleStoreShape,
-} from './postgres'
+} from './model'
+export { CycleStoreLive } from './postgres'
